add:Game Setting divide
